fix review score field

+++ api_yamdb/reviews/models.py
@@ -51,7 +51,7 @@ class Review(models.Model):
         Title, on_delete=models.CASCADE, related_name='reviews')
     text = models.TextField()
     # author = models.ForeignKey(User, on_delete=models.CASCADE, related_name='reviews')
-    score = models.FloatField()
+    score = models.IntegerField()
     pub_date = models.DateTimeField(auto_now_add=True)
